change PureComponent to Component, interface IDefaultSettings
